+++ crates/air_r_formatter/src/lib.rs
@@ -22,6 +22,8 @@ pub mod comments;
 pub mod context;
 mod cst;
 pub mod either;
+pub mod formatter_ext;
+pub mod joiner_ext;
 mod prelude;
 mod r;
 pub(crate) mod separated;