+++ crates/air_r_formatter/tests/specs/r/directives/skip-file-not.R
@@ -4,7 +4,7 @@ NULL
 
 # fmt: skip file
 
-1 + 1
+1+1