+++ crates/air_r_formatter/tests/specs/r/program.R
@@ -1,0 +1,42 @@
+
+
+
+
+# Up to two lines between elements
+1
+
+
+
+# Up to two lines between elements
+
+2
+
+
+
+# Up to two lines between elements
+
+fn <- function() {
+    1
+
+
+    # Up to one line between elements
+    2
+}
+
+
+# Up to two lines between elements
+# Currently failing but that doesn't seem too bad: It makes it so that a comment
+# is required to stick to its node with at most one empty line.
+
+
+
+1
+
+
+
+# Up to two lines between elements.
+# Currently failing but a trailing comment in the file does not matter too much.
+# To change the behaviour of these last two examples we'd have to fully take
+# over formatting of trivia in expression list
+
+