+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # Development version
 
+- We now allow up to 2 lines between top-level elements of a file. This makes it possible to separate long scripts into visually distinct sections (#40).
+
 - Unary formulas (i.e. anonymous functions) like `~ .x + 1` now add a space between the `~` and the right-hand side, unless the right-hand side is very simple, like `~foo` or `~1` (#235).
 
 - Semicolons at the very start or very end of a file no longer cause the parser to panic (#238).