+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # Development version
 
+- Language server configuration variables are now fully optional, avoiding issues in editors like Zed or Helix (#246).
+
 
 # 0.4.0